+++ CHANGELOG.md
@@ -7,6 +7,7 @@ See [release calendar](https://www.graalvm.org/release-calendar/) for release da
 
 ## Version 25.3.4
 * Implemented the [`Iterator Includes`](https://github.com/tc39/proposal-iterator-includes) proposal. It is available in ECMAScript staging mode (`--js.ecmascript-version=staging`).
+* Implemented the [`Iterator Join`](https://github.com/tc39/proposal-iterator-join) proposal. It is available in ECMAScript staging mode (`--js.ecmascript-version=staging`).
 
 ## Version 25.2.4
 * Updated Node.js to version 24.17.0.